Rename src/Controllers to src/Pages and move all Page-stuff there
