+++ _config/legacy.yml
@@ -3,7 +3,7 @@ Name: memberprofileslegacy
 ---
 SilverStripe\ORM\DatabaseAdmin:
   classname_value_remapping:
-    MemberProfilePage: Symbiote\MemberProfiles\Model\MemberProfilePage
+    MemberProfilePage: Symbiote\MemberProfiles\Pages\MemberProfilePage
     MemberProfileSection: Symbiote\MemberProfiles\Model\MemberProfileSection
     MemberProfileFieldsSection: Symbiote\MemberProfiles\Model\MemberProfileFieldsSection
     MemberProfileField: Symbiote\MemberProfiles\Model\MemberProfileField
