+++ _config/routes.yml
@@ -4,4 +4,4 @@ After: framework/routes#coreroutes
 ---
 SilverStripe\Control\Director:
   rules:
-    member-approval: Symbiote\MemberProfiles\Controllers\MemberApprovalController
+    member-approval: Symbiote\MemberProfiles\Pages\MemberApprovalController
